KVSystemServer/requirements.txt
Flask==3.0.0
aiohttp==3.9.3
websockets==12.0
python-multipart==0.0.9
Jinja2==3.1.2
supabase==2.3.0
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
asyncio==3.4.3
typing-extensions==4.9.0
urllib3==2.1.0
requests==2.31.0
flask-cors==4.0.0
